Extract extensions to separate file and remove debug statements
